Fix phpstan issues on ajax directory (#714)
* Fix phpstan issues on ajax directory

* Ajax files that seems no longer used, and rely on non existing functions/methods

* Fix CS

* Remove code that does not work

+++ ajax/dropdowntypelist.php
@@ -46,6 +46,6 @@ if ($myamemore != '') {
         $myamemore,
         filter_input(INPUT_POST, "method"),
         filter_input(INPUT_POST, filter_input(INPUT_POST, "name") . 'typeid'),
-        filter_input(INPUT_POST, "taskjobs_id")
+        (int) filter_input(INPUT_POST, "taskjobs_id")
     );
 }
