+++ ajax/taskjobaddtype.php
@@ -42,6 +42,6 @@ $pfTaskjob = new PluginGlpiinventoryTaskjob();
 $pfTaskjob->additemtodefatc(
     filter_input(INPUT_POST, "type"),
     filter_input(INPUT_POST, "itemtype"),
-    filter_input(INPUT_POST, "items_id"),
-    filter_input(INPUT_POST, "taskjobs_id")
+    (int) filter_input(INPUT_POST, "items_id"),
+    (int) filter_input(INPUT_POST, "taskjobs_id")
 );
